+++ react/ui-components/src/atoms/stories/BackLink.stories.js
@@ -6,33 +6,22 @@ export default {
   title: "Atoms/BackLink",
   component: BackLink,
   argTypes: {
-    variant: {
-      control: "select",
-      options: ["primary", "secondary", "teritiary"],
-    },
-    className: {
-      control: "text",
-    },
-    style: {
-      control: { type: "object" },
-    },
-    onClick: {
-      control: "function",
-    },
+    variant: { table: { disable: true } },
+    className: { table: { disable: true } },
+    style: { table: { disable: true } },
+    label: { table: { disable: true } },
+    onClick: { table: { disable: true } },
+    hideIcon: { table: { disable: true } },
+    hideLabel: { table: { disable: true } },
+    iconFill: { table: { disable: true } },
     disabled: {
-      control: "boolean",
-    },
-    hideLabel: {
-      control: "boolean",
-    },
-    hideIcon: {
-      control: "boolean",
-    },
-    iconFill: {
-      control: "text",
-    },
-    label: {
-      control: "text",
+      control: "select",
+      options: ["Default", "Disabled"],
+      name:"state",
+      mapping: {
+        Default: false, 
+        Disabled: true, 
+      },
     },
   },
 };
@@ -40,10 +29,9 @@ export default {
 const Template = (args) => <BackLink {...args} />;
 
 const commonArgs = {
-  className: "",
   style: {},
   onClick: () => console.log("clicked"),
-  disabled: false,
+  disabled: "Default",
   variant: "",
   hideIcon: false,
   hideLabel: false,
@@ -77,11 +65,4 @@ export const Backlink3 = Template.bind({});
 Backlink3.args = {
   ...commonArgs,
   variant: "teritiary",
-};
-
-export const Disabled = Template.bind({});
-Disabled.args = {
-  ...commonArgs,
-  variant: "primary",
-  disabled: true,
 };