+++ react/ui-components/src/atoms/stories/Tag.stories.js
@@ -8,49 +8,98 @@ export default {
   argTypes: {
     className: {
       control: "text",
+      table: { disable: true },
     },
     iconClassName: {
       control: "text",
+      table: { disable: true },
     },
     label: {
       control: "text",
+      name: "Label",
     },
     style: {
       control: { type: "object" },
+      table: { disable: true },
     },
     labelStyle: {
       control: { type: "object" },
+      table: { disable: true },
     },
     stroke: {
-      control: "boolean",
+      control: "check",
+      name: "Stroke",
+      options: ["Enable"],
     },
     type: {
       control: "select",
       options: ["monochrome", "success", "warning", "error"],
+      table: { disable: true },
     },
     alignment: {
       control: "select",
       options: ["center", "left", "right"],
+      table: { disable: true },
     },
     icon: {
       control: "text",
+      table: { disable: true },
     },
     showIcon: {
-      control: "boolean",
+      control: "check",
+      name: "Icon",
+      options: ["Enable"],
+    },
+    iconColor: {
+      table: { disable: true },
+    },
+    onClick: {
+      control: "check",
+      name: "Clickable",
+      options: ["Enable"],
     },
   },
 };
 
-const Template = (args) => <Tag {...args} />;
+const commonStyles = {
+  position: "absolute",
+  top: "50%",
+  left: "50%",
+  color: "#363636",
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "center",
+  transform: "translate(-50%, -50%)",
+};
+
+const Template = (args) => {
+  const { stroke, showIcon, onClick, ...rest } = args;
+  return (
+    <div style={commonStyles}>
+      <Tag
+        {...rest}
+        stroke={stroke?.length > 0 ? true : false}
+        showIcon={showIcon?.length > 0 ? true : false}
+        onClick={
+          onClick?.length > 0
+            ? (e) => {
+                console.log(e);
+              }
+            : undefined
+        }
+      />
+    </div>
+  );
+};
 
 const commonArgs = {
   label: "Tag",
   className: "",
   style: {},
-  stroke: false,
-  type: "success",
+  stroke: "",
+  type: "monochrome",
   icon: "",
-  showIcon: false,
+  showIcon: "",
   labelStyle: {},
 };
 
@@ -64,68 +113,36 @@ export const Documentation = () => (
 
 Documentation.storyName = "Docs";
 
-export const Default = Template.bind({});
-Default.args = {
+export const Basic = Template.bind({});
+Basic.args = {
   ...commonArgs,
+  type: "monochrome",
 };
 
-export const WithIcon = Template.bind({});
-WithIcon.args = {
+export const Success = Template.bind({});
+Success.args = {
   ...commonArgs,
-  label: "Tag With Icon",
-  showIcon: true,
+  type: "success",
 };
 
-export const WithStroke = Template.bind({});
-WithStroke.args = {
+export const Error = Template.bind({});
+Error.args = {
   ...commonArgs,
-  label: "Tag With Stroke",
-  stroke: true,
+  type: "error",
 };
-
-export const WithIconAndStroke = Template.bind({});
-WithIconAndStroke.args = {
+export const Warning = Template.bind({});
+Warning.args = {
   ...commonArgs,
-  label: "Tag With icon & stroke",
-  stroke: true,
-  showIcon: true,
+  type: "warning",
 };
-
-export const WithCustomIconAndStyles = Template.bind({});
-WithCustomIconAndStyles.args = {
+export const Custom = Template.bind({});
+Custom.args = {
   ...commonArgs,
-  label: "Tag With custom icons & colors",
-  stroke: true,
-  showIcon: true,
-  icon: "MyLocation",
-  style: {
-    backgroundColor: "#FCF2E4",
-  },
-  labelStyle: {
-    color: "#0B4B66",
-  },
-};
-
-export const WithOnClick = Template.bind({});
-WithOnClick.args = {
-  ...commonArgs,
-  label: "Tag With OnClick Function",
-  onClick: () => {
-    console.log("clicked");
-  },
-};
-
-export const WithAlignment = Template.bind({});
-WithAlignment.args = {
-  ...commonArgs,
-  label: "Tag With custom icons & colors",
-  stroke: true,
-  showIcon: true,
-  icon: "MyLocation",
   style: {
     width: "600px",
   },
   labelStyle: {
     color: "#0B4B66",
   },
+  icon: "ArrowLeft",
 };