+++ react/ui-components/src/molecules/stories/Footer.stories.js
@@ -1,47 +1,67 @@
 import React from "react";
 import { Button } from "../../atoms";
-import { ActionBar } from "../../atoms";
-import {Iframe} from "../../atoms";
+import { Footer } from "../../atoms";
+import { Iframe } from "../../atoms";
 
 export default {
   title: "Molecules/Footer",
-  component: ActionBar,
+  component: Footer,
   argTypes: {
     className: {
       control: "text",
+      table: { disable: true },
     },
     style: {
       control: { type: "object" },
+      table: { disable: true },
     },
     actionFields: {
       control: { type: "object" },
+      table: { disable: true },
     },
     setactionFieldsToRight: {
       control: { type: "boolean" },
+      table: { disable: true },
     },
     setactionFieldsToLeft: {
       control: { type: "boolean" },
+      table: { disable: true },
     },
     maxActionFieldsAllowed: {
       control: { type: "number" },
+      table: { disable: true },
     },
     sortActionFields: {
       control: { type: "boolean" },
+      table: { disable: true },
     },
     setChildrenLeft: {
       control: { type: "boolean" },
+      table: { disable: true },
+    },
+    flex: {
+      control: "select",
+      options: ["Left", "Right"],
+      name:"Flex"
     },
   },
 };
 
-const Template = (args) => <ActionBar {...args} />;
+const Template = (args) => {
+  const { flex, ...rest } = args;
+  return (
+    <Footer
+      {...rest}
+      setactionFieldsToRight={flex === "Right"}
+      setactionFieldsToLeft={"Left"}
+    />
+  );
+};
 
 const commonArgs = {
   className: "",
   style: {},
   actionFields: [],
-  setactionFieldsToRight: false,
-  setactionFieldsToLeft: false,
   maxActionFieldsAllowed: 5,
   sortActionFields: true,
   setChildrenLeft: false,
@@ -87,31 +107,9 @@ const footeractionFieldsWithSearchableDropdown = [
     variation={"primary"}
     optionsKey={"name"}
     isSearchable={true}
-    onOptionSelect={(option)=>{console.log(option)}}
-  ></Button>,
-];
-
-const footeractionFieldsWithUnSearchableDropdown = [
-  <Button
-    type={"button"}
-    size={"large"}
-    variation={"secondary"}
-    label="Back"
-    icon={"ArrowBack"}
-    onClick={() => console.log("Moving to previous page")}
-  />,
-  <Button
-    type={"actionButton"}
-    options={[
-      { name: "Action A", code: "Actiona" },
-      { name: "Action B", code: "Actionb" },
-      { name: "Action C", code: "Actionc" },
-    ]}
-    label={"Actions"}
-    variation={"primary"}
-    optionsKey={"name"}
-    isSearchable={false}
-    onOptionSelect={(option)=>{console.log(option)}}
+    onOptionSelect={(option) => {
+      console.log(option);
+    }}
   ></Button>,
 ];
 
@@ -125,35 +123,23 @@ export const Documentation = () => (
 
 Documentation.storyName = "Docs";
 
+export const Basic = () => (
+  <Footer
+    className={""}
+    style={{}}
+    actionFields={footeractionFieldsWithSearchableDropdown}
+    maxActionFieldsAllowed={5}
+    sortActionFields={true}
+    setChildrenLeft={false}
+  />
+);
+Basic.argTypes = {
+  flex: { table: { disable: true } },
+};
 
-export const Default = Template.bind({});
-Default.args = {
+export const Flex = Template.bind({});
+Flex.args = {
   ...commonArgs,
   actionFields: footeractionFields,
-};
-
-export const actionFieldsSetToRight = Template.bind({});
-actionFieldsSetToRight.args = {
-  ...commonArgs,
-  actionFields: footeractionFields,
-  setactionFieldsToRight: true,
-};
-
-export const actionFieldsSetToLeft = Template.bind({});
-actionFieldsSetToLeft.args = {
-  ...commonArgs,
-  actionFields: footeractionFields,
-  setactionFieldsToLeft: true,
-};
-
-export const WithSearchableDropdown = Template.bind({});
-WithSearchableDropdown.args = {
-  ...commonArgs,
-  actionFields: footeractionFieldsWithSearchableDropdown,
-};
-
-export const WithUnsearchableDropdown = Template.bind({});
-WithUnsearchableDropdown.args = {
-  ...commonArgs,
-  actionFields: footeractionFieldsWithUnSearchableDropdown,
+  flex:"Right"
 };